+++ src/main/java/org/dependencytrack/model/Vulnerability.java
@@ -322,8 +322,8 @@ public class Vulnerability implements Serializable {
 
     @Persistent(table = "VULNERABILITY_TAGS", defaultFetchGroup = "true")
     @Join(column = "VULNERABILITY_ID")
-    @Element(column = "VULNERABILITY_TAG_ID")
-    private List<VulnerabilityTag> vulnerabilityTags;
+    @Element(column = "TAG_ID")
+    private List<Tag> tags;
 
     private transient Epss epss;
 
@@ -741,11 +741,11 @@ public class Vulnerability implements Serializable {
         return epss != null ? epss.getPercentile() : null;
     }
 
-    public List<VulnerabilityTag> getVulnerabilityTags() {
-        return vulnerabilityTags;
+    public List<Tag> getTags() {
+        return tags;
     }
 
-    public void setVulnerabilityTags(List<VulnerabilityTag> vulnerabilityTags) {
-        this.vulnerabilityTags = vulnerabilityTags;
+    public void setTags(List<Tag> tags) {
+        this.tags = tags;
     }
 }
