+++ src/main/java/org/dependencytrack/persistence/VulnerabilityQueryManager.java
@@ -29,6 +29,7 @@ import org.dependencytrack.model.ConfigPropertyConstants;
 import org.dependencytrack.model.Epss;
 import org.dependencytrack.model.FindingAttribution;
 import org.dependencytrack.model.Project;
+import org.dependencytrack.model.Tag;
 import org.dependencytrack.model.Vulnerability;
 import org.dependencytrack.model.VulnerabilityAlias;
 import org.dependencytrack.model.VulnerableSoftware;
@@ -886,5 +887,31 @@ final class VulnerabilityQueryManager extends QueryManager implements IQueryMana
         query.setFilter("vulnerability == :vulnerability");
         query.setParameters(vulnerability);
         query.deletePersistentAll();
+    }
+
+    /**
+     * Binds the two objects together in a corresponding join table.
+     *
+     * @param vulnerability a Vulnerability object
+     * @param tags    a List of Tag objects
+     */
+    @SuppressWarnings("unchecked")
+    public void bind(Vulnerability vulnerability, List<Tag> tags) {
+        final Query<Tag> query = pm.newQuery(Tag.class, "vulnerabilities.contains(:vulnerability)");
+        final List<Tag> currentVulnerabilityTags = (List<Tag>) query.execute(vulnerability);
+        pm.currentTransaction().begin();
+        for (final Tag tag : currentVulnerabilityTags) {
+            if (!tags.contains(tag)) {
+                tag.getVulnerabilities().remove(vulnerability);
+            }
+        }
+        vulnerability.setTags(tags);
+        for (final Tag tag : tags) {
+            final List<Vulnerability> vulnerabilities = tag.getVulnerabilities();
+            if (!vulnerabilities.contains(vulnerability)) {
+                vulnerabilities.add(vulnerability);
+            }
+        }
+        pm.currentTransaction().commit();
     }
 }
