+++ src/main/java/org/dependencytrack/resources/v1/VulnerabilityResource.java
@@ -34,8 +34,8 @@ import org.dependencytrack.model.AnalyzerIdentity;
 import org.dependencytrack.model.Component;
 import org.dependencytrack.model.Cwe;
 import org.dependencytrack.model.Project;
+import org.dependencytrack.model.Tag;
 import org.dependencytrack.model.Vulnerability;
-import org.dependencytrack.model.VulnerabilityTag;
 import org.dependencytrack.model.VulnerableSoftware;
 import org.dependencytrack.persistence.QueryManager;
 import org.dependencytrack.resources.v1.vo.AffectedComponent;
@@ -373,9 +373,9 @@ public class VulnerabilityResource extends AlpineResource {
                     }
                 }
 
-                final List<VulnerabilityTag> resolvedTags = qm.resolveVulnerabilityTags(jsonVuln.getVulnerabilityTags());
-                qm.bindTagsToVulnerability(vulnerability, resolvedTags);
-                
+                final List<Tag> resolvedTags = qm.resolveTags(jsonVuln.getTags());
+                qm.bind(vulnerability, resolvedTags);
+
                 recalculateScoresFromVector(jsonVuln);
                 vulnerability = qm.updateVulnerability(jsonVuln, true);
                 qm.persist(vsList);
